Refactor CPU and MMU for CGB speed switching support
